refactor: API 요청 수정
